Move Faucet dApp to plugins
